requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.31.0
tensorflow>=2.0.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
